+++ test-resources/lib_tests/failjure/runner.cljs
@@ -1,0 +1,5 @@
+(ns failjure.runner
+    (:require [doo.runner :refer-macros [doo-tests]]
+              [failjure.test-core]))
+
+(doo-tests 'failjure.test-core)