+++ test-resources/lib_tests/markdown/runner.cljs
@@ -1,0 +1,5 @@
+(ns markdown.runner
+  (:require [doo.runner :refer-macros [doo-tests]]
+            [markdown.md-test]))
+
+(doo-tests 'markdown.md-test)