+++ test-resources/lib_tests/medley/test_runner.cljs
@@ -1,0 +1,5 @@
+(ns medley.test-runner
+  (:require [doo.runner :refer-macros [doo-tests]]
+            [medley.core-test]))
+
+(doo-tests 'medley.core-test)