+++ test-resources/lib_tests/table/core_test.clj
@@ -15,7 +15,7 @@
       +---+---+
       | 3 | 4 |
       +---+---+
-      ") "\n")
+      ") (System/lineSeparator))
     (with-out-str (table [["1" "2"] ["3" "4"]])))))
 
 (deftest test-table-with-vecs-in-vec
@@ -310,7 +310,7 @@
       ")
     (table-str  [[1 2] [:c :d]  [:a :b]] :sort true))))
 
-;; BB-TEST-PATCH: Intermittent failing test
+;; TODO: Intermittent failing test
 #_(deftest test-table-with-sort-option-as-field-name
   (is (=
     (unindent